clean the code after review
