+++ huntflow_api_client/api.py
@@ -1,12 +1,14 @@
-import time
-from typing import Any, Dict, Callable, List, Optional
+import logging
+from typing import Callable, List, Optional
 
 import httpx
 
 from huntflow_api_client.errors import TokenExpiredError, InvalidAccessTokenError
-from huntflow_api_client.api_token.base_token_proxy import AbstractTokenProxy
-from huntflow_api_client.api_token.huntflow_token_proxy import DummyHuntflowTokenProxy
-from huntflow_api_client.api_token.huntflow_token import HuntflowApiToken
+from huntflow_api_client.tokens.proxy import AbstractTokenProxy, DummyHuntflowTokenProxy
+from huntflow_api_client.tokens.token import HuntflowApiToken
+
+
+logger = logging.getLogger(__name__)
 
 
 class HuntflowAPI:
@@ -20,6 +22,23 @@ class HuntflowAPI:
         request_event_hooks: Optional[List[Callable]] = None,
         response_event_hooks: Optional[List[Callable]] = None,
     ):
+        """API client.
+        :param base_url: Base url for API, like 'https://<api domain>/v2'
+        :param token: Optional token structure with access token.
+            Use it if you don't care about token refreshing
+        :param token_proxy: Alternative way (see `token` param) to provide token.
+            If you need to properly refresh tokens, then you should use this param
+            instead of `token`. You can use `huntflow_api_client.tokens.proxy.HuntflowTokenProxy`
+            or implement your own class for this,
+            the class must implement AbstractTokenProxy interface.
+            Also see usage example at `examples.api_client_with_simple_locks`.
+        :param auto_refresh_tokens: If True then the client will handle token expiration.
+            "Handle" means: catch token expiration errors and run token refresh request.
+        :param request_event_hooks: callable to run before actual requests.
+        :param response_event_hooks: callable to run after requests. Use it carefully
+            if you need auto-token refresh. If it hides (or modifies in some way) 401
+            errors, then, most likely, token refresh doesn't work.
+        """
         if token_proxy is None:
             if token is None:
                 raise ValueError("Parameters token and token_proxy can't be None at the same time")
@@ -121,13 +140,16 @@ class HuntflowAPI:
         # Consider the situation:
         # * we send 4 requests at the same time
         # * 3 of the 4 are returned with 401 token_expired errors
-        # * 1 of the 4 is still waiting for response (any reasons, may be rate limit at API server side)
-        # * one of failed 3 requests have refreshed the token, the rest of the 3 will use the refreshed
-        #   token for retries
-        # * the waiting request have got 401 token_expired response. But the token has been just refreshed,
-        #   so there is no need to refresh it again.
-        #   To track this case we have to check if the token has been updated (since last get_auth_header call).
-        #   If token has been updated, then we just need to retry original request with a refreshed auth data.
+        # * 1 of the 4 is still waiting for response
+        #   (any reasons, may be rate limit at API server side)
+        # * one of failed 3 requests have refreshed the token,
+        #   the rest of the 3 will use the refreshed token for retries
+        # * the waiting request have got 401 token_expired response.
+        #   But the token has been just refreshed, so there is no need to refresh it again.
+        #   To track this case we have to check if the token has been updated
+        #   (since last get_auth_header call).
+        #   If token has been updated, then we just need to retry original request with a
+        #   refreshed auth data.
         if await self._token_proxy.is_updated():
             return
         if not await self._token_proxy.lock_for_update():
